Update object detection files
