--- image/object-detection/preproc_combined.py
@@ -1,24 +1,0 @@
-import cv2
-import numpy as np
-from torchvision import transforms
-
-
-def preprocess_function(img: np.ndarray, bboxes: np.ndarray) -> np.ndarray:
-    '''Preprocess function used for YOLOX models'''
-    input_size = (416, 416)
-    if len(img.shape) == 3:
-        padded_img = np.ones((input_size[0], input_size[1], 3), dtype=np.uint8) * 114
-    else:
-        padded_img = np.ones(input_size, dtype=np.uint8) * 114
-
-    r = min(input_size[0] / img.shape[0], input_size[1] / img.shape[1])
-    resized_img = cv2.resize(
-        img,
-        (int(img.shape[1] * r), int(img.shape[0] * r)),
-        interpolation=cv2.INTER_LINEAR,
-    ).astype(np.uint8)
-    padded_img[: int(img.shape[0] * r), : int(img.shape[1] * r)] = resized_img
-
-    padded_img = padded_img.transpose((2, 0, 1))
-    padded_img = np.ascontiguousarray(padded_img, dtype=np.float32)
-    return padded_img, bboxes * r
